requirements.txt
llvmlite==0.40.1
numba==0.57.1
numpy==1.24.2
Pillow==10.2.0
PyOpenGL==3.1.6
PySide6==6.2.4
shiboken6==6.2.4
scipy==1.11.2